+++ sdk/src/lib.rs
@@ -211,9 +211,8 @@ pub mod crypto;
 #[doc(hidden)]
 pub mod dynamic_assertion;
 
-// TODO: pub it when we expose in high-level API
 /// The `http` module contains generic traits for configuring sync and async HTTP resolvers.
-pub(crate) mod http;
+pub mod http;
 
 /// The `identity` module provides support for the [CAWG identity assertion](https://cawg.io/identity).
 #[doc(hidden)]
@@ -266,10 +265,7 @@ pub use resource_store::{ResourceRef, ResourceStore};
 #[doc(inline)]
 pub use settings::Settings;
 pub use signer::{AsyncSigner, BoxedAsyncSigner, BoxedSigner, Signer};
-#[cfg(any(not(target_arch = "wasm32"), target_os = "wasi"))]
-#[cfg_attr(docsrs, doc(cfg(any(not(target_arch = "wasm32"), target_os = "wasi"))))]
-pub use utils::ephemeral_signer::EphemeralSigner;
-pub use utils::mime::format_from_path;
+pub use utils::{ephemeral_signer::EphemeralSigner, mime::format_from_path};
 #[doc(inline)]
 pub use validation_results::{ValidationResults, ValidationState};