+++ c2pa_c_ffi/CHANGELOG.md
@@ -7,6 +7,34 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.77.1](https://github.com/contentauth/c2pa-rs/compare/c2pa-c-ffi-v0.77.0...c2pa-c-ffi-v0.77.1)
+_10 March 2026_
+
+### Fixed
+
+* Add API to read manifest bytes using reader+context ([#1904](https://github.com/contentauth/c2pa-rs/pull/1904))
+* Return -1 on settings failure ([#1905](https://github.com/contentauth/c2pa-rs/pull/1905))
+
+## [0.77.0](https://github.com/contentauth/c2pa-rs/compare/c2pa-c-ffi-v0.76.2...c2pa-c-ffi-v0.77.0)
+_04 March 2026_
+
+### Added
+
+* Embeddable manifest API with Context/Settings, CAWG and BMFF.v3 support. ([#1847](https://github.com/contentauth/c2pa-rs/pull/1847))
+
+## [0.76.2](https://github.com/contentauth/c2pa-rs/compare/c2pa-c-ffi-v0.76.1...c2pa-c-ffi-v0.76.2)
+_03 March 2026_
+
+### Added
+
+* Signer programatically configured on Context (C FFI APIs) ([#1893](https://github.com/contentauth/c2pa-rs/pull/1893))
+
+## [0.76.1](https://github.com/contentauth/c2pa-rs/compare/c2pa-c-ffi-v0.76.0...c2pa-c-ffi-v0.76.1)
+_02 March 2026_
+
+## [0.75.22](https://github.com/contentauth/c2pa-rs/compare/c2pa-c-ffi-v0.75.21...c2pa-c-ffi-v0.75.22)
+_23 February 2026_
+
 ## [0.75.21](https://github.com/contentauth/c2pa-rs/compare/c2pa-c-ffi-v0.75.20...c2pa-c-ffi-v0.75.21)
 _12 February 2026_